ci(gtpu-dataplane): gate the classifier load on the OpenShift 4.18 kernel
OpenShift 4.18 runs RHCOS on RHEL 9.4, so its nodes are the 5.14.0 line
rather than the 6.8 mainline kernel CI proved. That line is an
enterprise backport: it carries bpf_loop, but helper presence does not
establish that this object's callback chain fits the verifier's
cumulative 512-byte stack accounting, which is the only property that
decides whether the datapath loads. Stating the line as unqualified was
therefore the wrong answer for a supported deployment target.

Gate it instead. A second QEMU job boots a Rocky 9 guest -- ABI
equivalent to RHEL 9 and publicly downloadable, unlike RHEL itself --
asserts the guest really is on 5.14 before trusting anything, and loads
both committed classifiers plus the capability probe there.

The blocker was binary portability, not the kernel: the test binary is
built on an Ubuntu runner and executed unmodified in the guest, and a
gnu build will not start against el9's glibc. Building for
x86_64-unknown-linux-musl produces a static binary that runs on either,
and the job asserts it is statically linked so a silent regression to a
dynamic build cannot pass.

That build needed one portability fix: statfs::f_type is signed on musl
and unsigned on glibc, so the bpffs magic comparison now widens both
sides instead of assuming one libc.

Both READMEs now record 5.14/el9 as a gated line rather than an
unqualified one, and keep the distinction the issue asked for: helper
availability is not loadability, and kernels outside the gated lines
still have no verdict.

Refs #543

+++ crates/opc-linux-xfrm-sys/src/linux.rs
@@ -121,7 +121,9 @@ fn verify_bpffs(fd: &OwnedFd) -> io::Result<()> {
     if unsafe { libc::fstatfs(fd.as_raw_fd(), &mut status) } < 0 {
         return Err(io::Error::last_os_error());
     }
-    if status.f_type != BPF_FS_MAGIC {
+    // `statfs::f_type` is signed on musl and unsigned on glibc, so widen both
+    // sides into a type that holds either rather than assuming one libc.
+    if i128::from(status.f_type) != i128::from(BPF_FS_MAGIC) {
         return Err(io::Error::new(
             io::ErrorKind::InvalidInput,
             "pin directory is not on bpffs",
